+++ solc_select/models/__init__.py
@@ -1,0 +1,5 @@
+"""
+Domain models for solc-select.
+
+This package contains the core business entities and value objects organized by concern.
+"""
