+++ solc_select/services/repository_matcher.py
@@ -8,7 +8,9 @@ repository for a requested version based on platform capabilities and repository
 import requests
 
 from ..exceptions import VersionNotFoundError
-from ..models import Platform, RepositoryManifest, SolcVersion
+from ..models.platforms import Platform
+from ..models.repositories import RepositoryManifest
+from ..models.versions import SolcVersion
 from ..platform_capabilities import PlatformCapability, PlatformIdentifier
 from ..repositories import (
     AlloyRepository,
